restructured files schema files to model
