+++ vic-class11-assignment/gulpfile.js
@@ -1,11 +1,10 @@
 const gulp = require('gulp');
 const eslint = require('gulp-eslint');
 const mocha = require('gulp-mocha');
-const watch = require('gulp-watch');
 
 gulp.task('default', ['lint', 'test'], () => {
   console.log('started');
-})
+});
 
 gulp.task('lint', () => {
   gulp.src('/server.js')
