+++ vic-class11-assignment/model/icecream.js
