+++ vic-class11-assignment/package.json
@@ -27,7 +27,9 @@
     "mocha": "^2.5.2"
   },
   "dependencies": {
+    "bcrypt": "^0.8.6",
     "body-parser": "^1.15.1",
-    "express": "^4.13.4"
+    "express": "^4.13.4",
+    "jsonwebtoken": "^7.0.0"
   }
 }
