--- vic-class11-assignment/schema/user.js
@@ -1,10 +1,0 @@
-'use strict';
-
-const mongoose = require('mongoose');
-
-const User = new mongoose.Schema({
-  username: {type: String, required: true},
-  password: {type: String, required: true}
-});
-
-module.exports = mongoose.model('user', User);
